add help button to  home page
