+++ client/src/components/buttons/admin.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { Button } from 'antd';
+import { useNavigate } from 'react-router-dom';
+
+export default function AdminButton({ style }) {
+  const navigate = useNavigate();
+  return (
+    <Button
+      type='default'
+      shape='round'
+      onClick={() => navigate('/admin')}
+      style={style}
+    >
+      Admin
+    </Button>
+  );
+}
